Removed signin, signout, and login files and integrated everything onto the landing page
