+++ src/routes/+layout.server.ts
@@ -1,8 +1,17 @@
-export const load = async ({ locals }) => {
-    const session = await locals.getSession();
-    const loggedIn = !!session?.user;
+// export const load = async ({ locals }) => {
+//     const session = await locals.getSession();
+//     const loggedIn = !!session?.user;
   
-    return {
-      loggedIn,
-    };
-  };
+//     return {
+//       loggedIn,
+//     };
+//   };
+import type { LayoutServerLoad } from "./$types"
+
+// Loads serverside any data we need
+export const load: LayoutServerLoad = async (event) => {
+  return {
+    // Our serverside handle is adding data to event.locals for us
+    session: await event.locals.auth()
+  }
+}
